NF: move Preferences.kt to anki.preferences package
