+++ AnkiDroid/src/main/java/com/ichi2/anki/preferences/ReviewingSettingsFragment.kt
@@ -17,7 +17,6 @@ package com.ichi2.anki.preferences
 
 import androidx.preference.ListPreference
 import androidx.preference.SwitchPreference
-import com.ichi2.anki.Preferences
 import com.ichi2.anki.R
 import com.ichi2.anki.reviewer.AutomaticAnswerAction
 import com.ichi2.libanki.backend.exception.BackendNotSupportedException
